Search for eslint in parent directories and fix `this.binary` caching
This tells linter-eslint to recursively search parent directories for the eslint binary. This supports projects that share a eslint configuration between multiple subprojects:

```
.
├── Client
│   ├── client.js
│   └── package.json
├── Server
│   ├── server.js
│   └── package.json
└── package.json  # eslint, babel-eslint, eslint-plugins here
```

With this commit, if you open up the Client or Server directory in Atom, linter-eslint will find eslint in the parent directory's node_modules.

There was also a bug where `this.binary` would add a property to the global object instead of the exports, since `findESLint` was using unbound functions. Switched to arrow functions to fix this.