Pass eslint relative paths of files if possible
Fixes #293

.eslintignore files must be placed at a project's root, and the patterns
within the .eslintignore are relative to project root.  Therefore, for ESLint
to know whether a file should be ignored, one of two things must be true:

1. The .eslintignore entries need to be in the format `**/foo.js`
2. The paths passed to eslint for `--stdin-filename` need to be relative to
the project root, which is where the .eslintignore must be located.

This change will check for the existence of a `.eslintignore` file, and if one
is found, the arguments to `--stdin-filename` will be relative to its directory.