Provide correct filename to eslint
By default, process.cwd is the directory containing the file being linted, so
normally it is only necessary to provide the name of the file itself, with no path.

But, if a .gitignore file is present, eslint will not properly ignore a file if
cwd is not set to the project root when executing with text and providing
a filename, as we are doing. So, if we find an .eslintignore, we can simply
change the cwd to the root directory (same directory as the .eslintignore).