Test from root if absolute path is given
Without this change, when an absolute path is provided, the leading slash
is used as a separator and the first element of the chunks array is an 
empty string.  

With this change, the first element is replaced with a path separator,
so that the reconstructed currentDirectory is an absolute path, as desired.