add link to eslint rule in message
User can now click the link in the linter message to see the rule
definition on the eslint site.

Also added a missing space between the rule span and the rule message.

The only side-effect I'm aware of is that css for my theme styles links
close in color to the badge, as seen in this screenshot:
http://cl.ly/2O2o0Y2o0O3L. However, the rule at the bottom looks good,
again in my theme (One Dark).