Add config option to specify a custom eslint installation directory
This commit adds a new config option to specify a custom eslint
installation directory, which can be a path relative to the atom
project, or a global path.
If the option is provided, it takes precendence over the global option,
and will be used as a first option instead of a global or local installations.

Adds specs and fixes a test that was incorrectly passing