Fix specs for ‘fix’ functionality
The spec was previously attempting to test the fix-on-save functionality,
but unfortunately due to the way `lint` is set up in the specs, the 
logic was never exercised which would have applied fixing after a save.

So, the file was being linted, saved without changing, and then there
were some errors thrown which prevented all of the assertions from being
hit.

This change simplifies the test to only exercise the `fix` type of worker
job.  It does this by:

1. Copying the fixture file and a config file to a temporary directory
2. Linting the file to ensure it has a few errors
3. Setting a watcher on the file which will detect when it changes and
assert that there are no longer errors
4. Performing a fix on the file, which should change the file and kick off
the watcher.
5. Removing the copied files and temp directory