Move to babel-cli
Replace the current `ucompiler` setup with a standard `babel-cli` setup,
along with a configuration that transpiles _only_ what is necessary,
instead of the current method of transpiling everything that _might_ be
needed. This currently transpiles to support for Node.js v5 as that is
what Atom v1.10.0 ships with, which is our current minimum version.