Determine fix results based on exit code
Perhaps in the past ESLint threw an error if `--fix` did not fix all of
the errors in a file.  Or, maybe this logic was flawed from the start.
Either way, this change just looks at the exit code (`0` if there are 
no linting errors left, `1` if there are), and returns an appropriate
string.  Both cases are really success messages, there were no errors.

If there was an error for some reason, we need to print the `err.message`
string, because `addWarning` expects a string, not an error object.