Remove custom styling
The forced styling being done before was breaking on several themes, to
the point where some were unreadable. This removes the badge styling and
instead lets the theme handle the URL as a simple link. Rule name is
moved to the end of the message, similar to ESLint's CLI output.