Find closest package.json rather than home .eslintrc
Fixes #832

Previously, we would find the fallback `.eslintrc.*` config file in a
user’s home directory, return that, and then potentially ignore it later
because it’s a fallback.  Only if there were _zero_ `.eslintrc.*` files
would we ever look for a package.json.

With this change, we will find a package.json file and check for an 
eslint configuration before traversing higher up the directory structure
and potentially encountering a fallback in the user’s home root.

It’s important that the `package.json` file come last in the array, so
that we look for `.eslintrc` config files first.  Otherwise, we may
find a package.json file without an eslintConfig, ignore it, and not bother
looking for a .eslintrc file in the same directory.

This still has the problem that hitting a `package.json` file without
an eslintConfig will prevent finding another config higher in the 
structure.  That fix will come in the next commit.