Lazy require all dependencies of the main file
Move to lazily requiring all dependencies of the main file so
"activation" time is kept down, as TimeCop includes this in the
activation instead of loading time.

Note that idsToIgnoredRules was moved into main.js so we aren't forced
to require it to observe settings. Although this requires a
re-implementation in the specs, the specs already are containing some
implementation in there in the first place so I don't think this is a
big deal for such a small bit of logic.