Ensure the worker is started when needed
Since idle callbacks are handled in a FIFO queue if the worker has not
been initialized yet we can wait on another idle callback to ensure that
the idle callback that initializes the worker has been called.

Fixes a semi-rare case where a lint call happens before the worker has
been initialized.