Allow ESLint to fully determine config when possible
ESLint already handles grabbing the configuration for a file for us, but
we have actually been preventing this from working properly by always
specifying the path to what we determined to be the configuration file.

Since we really only need that to determine if we should run ESLint at
all, stop sending that to ESLint unless we are overriding it with a
fallback.