Wait on specific Notification in specs
As the specs execute asynchronously, it's possible for there to be more
than one Notification observer active at a time, leading to some specs
intermittently failing when they grab the "wrong" notification. Enhance
the function that waits on a notification to check for an expected
message as well, and silently discard the received notification if it
doesn't match what the spec is looking for.