Support global Yarn installs and better errors
Add support for global Yarn installations. Apparently one of the few
things Yarn did right is that all platforms use the same directory
structure under the base directory, unfortunately that meant that on
non-Windows platforms specifying the path still got you an unusable
setup.

Also fixes the error messages that can be thrown to better reflect the
actual causes of the messages.