Try to set cwd to project root dir if no .eslintignore
This tweaks our fallback cwd and relative path resolution slightly, such
that instead of falling straight back to just the directory of the 
file itself, we will use the base directory open in the atom editor.

If for whatever reason that is `null` (or undefined), we’ll still fall
back to the file directory as before.