Disable spying on Date.now by Jasmine
It turns out that `TextBuffer::onDidReload` is being broken _only_ in
the specs, as it relies on Date.now() which Jasmine was mocking. We can
fix this by telling Jasmine to stop spying on that function.

Also re-enables a long-disabled test that looks to be working properly
now that the TextBuffer rewrite has landed.