Add option to silence fixable rules while typing
Add an option that silences fixable rules while typing. The list of
fixable rules is populated from ESLint itself, and only sent back to the
main package code when it updates to reduce message size. This means
that this will only start silencing rules after the first lint rule.
Since a lint is triggered on open by default this _should_ be largely
transparent to the user.

Fixes #894.