Group disabling options together
This creates an options group for settings that disable linting
in one way or another.

Because moving options into groups will require a lot of migrations,
this also creates a new file/function to handle those migrations,
thus keeping them out of `main.js`.

There may be a more elegant way to manage the migrations, but this gets
us started.