Streamline config setting migrations
For standard migrations, all we really need to do is check if the old
setting is defined, then move it to the new setting, then unset the old
setting.  We can just make a simple loop for this, keeping the old
and new setting strings as data, apart from the logic that moves them.