Remove useless "if" wrappers around requires
Node's `require` stores imports in memory. If you require the same module
multiple times, it will reuse the first import each time. Using late requires is not ideal because `require` is blocking when it does need to load a file, but if we are going to use it, we can at least use it as designed.

Patch removes if checks around requires and replaces global let
variables with const variables local to the function that uses the
imports.