Implement an alternative test runner
As per https://github.com/AtomLinter/Meta/issues/19, it would be ideal to have a test runner with code coverage capabilities. This is a first try at getting Mocha or Jest running inside the Atom environment.

This commit has a working installation of mocha, with test showing the atom globals are found showed as passing in the Atom Spec Runner Window.

Jest is not working at this time. Jest will run inside Atom and try to test the file, but fails to inject the Jest global variables, such as `test` and `expect`.