Update: Store rules w/ redux, and diff rules
Adds a redux data-store, with reducer and actions for storing `rules` in
the store, and selectors for retrieving computed data. Adds diffing to
rules to reduce overhead of passing between threads. Includes tests for
new code, a few generic utility functions, and directory structure
preimptively laid out for dealing with more pieces of state in the
store.