Mv Atom config retrieval out of main
All settings are retrieved through observation except for migrated settings (which is run in idleCallback toreduce impact on typical already-upgraded users) and debug (which wants the raw settings). The config subscriptions and migrations are extracted to an atom-config submodule. The config passed to lint jobs is reduced to only properties that get used.