Organize config options (#1042)
* Remove unused docs images

* Remove mention of blog post

It’s out-of-date, and not actually even running right now.

* Rearrange and order settings

By default, Atom will sort settings alphabetically, which can result
in a really strange, non-logical order.  This commit attempts to put 
more important or commonly used settings up top, while also grouping
similar settings together.  So, the scopes are together with the lint
html setting, the use global and global paths are together, the fixing
options are grouped, etc.  I’m sure this isn’t perfect, but it’s a start.

* Group disabling options together

This creates an options group for settings that disable linting
in one way or another.

Because moving options into groups will require a lot of migrations,
this also creates a new file/function to handle those migrations,
thus keeping them out of `main.js`.

There may be a more elegant way to manage the migrations, but this gets
us started.

* Streamline config setting migrations

For standard migrations, all we really need to do is check if the old
setting is defined, then move it to the new setting, then unset the old
setting.  We can just make a simple loop for this, keeping the old
and new setting strings as data, apart from the logic that moves them.

* Group “advanced” settings

These are settings that are not commonly needed, so we may as well
try to avoid confusing new users by indicating that they are “advanced”

* Group global eslint options together

* Rearrange remaining options

* Wording tweaks

* Move `eslintrcPath` setting to `global` section

* Move eslintRulesDirs to advanced section

And also collapse advanced section.

* Create an `autofix` section

* Do a single `atom.config.get`

* Refine migration approach slightly

* Avoid object rest spread

(we don’t have babel configured for it)

* Rename advancedLocalNodeModules to localNodeModules

Since it has a prefix of `advanced` now anyways.

* Make “called” method of spy return a boolean

I believe this is what was intended originally.

* Fix migrate-config-options and add specs