Upgrade: Pin Dependencies
Pin the version of the dependencies that would have been installed via 
the old ranges if the package-lock.json wasn't present. Future updates 
to dependencies will ensure the package-lock.json stays consistent now, 
so we don't have things becoming out of date in the version that is 
getting installed without us realizing it.